move PressedCollectionVisibility into own file and add tests
